+++ wiki/Changelog.md
@@ -5,6 +5,7 @@ Terminal features
   * Fix (revert back) DECSDM (DECSET 80) Sixel Display mode (#1127, xterm 369).
   * Sound file playing OSC 440 (#1122).
   * DECPS tone playing support (#1122).
+  * Fixed LED state glitch when ScrollLock is held in auto-repeat.
 
 Window handling
   * Fixed -s max... options (#1124).